// src/generated/models/DCPO_LISTE_AGENCESModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface AuthorValue {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface EditorValue {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface DCPO_LISTE_AGENCESBase {
  ID?: number;
  Title?: string;
  field_1?: number;
  OData__ColorTag?: string;
  ComplianceAssetId?: string;
  Modified?: string;
  Created?: string;
  "Author#Claims"?: string;
  "Editor#Claims"?: string;
  "{Identifier}"?: string;
  "{IsFolder}"?: boolean;
  "{Thumbnail}"?: {
    "@odata.type"?: string;
    // URL vers une miniature présentant une version fortement compressée de l’élément, si cela est possible.
    Small?: string;
    // URL vers la miniature de taille standard de l’élément, le cas échéant.
    Medium?: string;
    // URL vers la miniature la plus grande de l’élément, le cas échéant.
    Large?: string;
  };
  "{Link}"?: string;
  "{Name}"?: string;
  "{FilenameWithExtension}"?: string;
  "{Path}"?: string;
  "{FullPath}"?: string;
  "{ModerationStatus}"?: string;
  "{ModerationComment}"?: string;
  "{ContentType}#Id"?: string;
  "{ContentType}"?: {
    "@odata.type"?: string;
    Id?: string;
    Name?: string;
  };
  "{HasAttachments}"?: boolean;
  "{Attachments}@odata.type"?: string;
  "{Attachments}"?: {
   "@odata.type"?: string;
   Id?: string;
   AbsoluteUri?: string;
   DisplayName?: string;
 }[];
  "{VersionNumber}"?: string;
  "{TriggerWindowStartToken}"?: string;
  "{TriggerWindowEndToken}"?: string;
}

export interface DCPO_LISTE_AGENCESRead extends DCPO_LISTE_AGENCESBase {
  Author?: AuthorValue;
  Editor?: EditorValue;
}

export interface DCPO_LISTE_AGENCESWrite extends DCPO_LISTE_AGENCESBase {
  Author?: AuthorValue;
  Editor?: EditorValue;
}